+++ src/Lakona.Game.Server/Actors/IActorActivationLifecycle.cs
@@ -1,0 +1,6 @@
+namespace Lakona.Game.Server.Actors;
+
+internal interface IActorActivationLifecycle
+{
+    ValueTask DrainAsync(CancellationToken cancellationToken = default);
+}